+++ rstar/src/algorithm/rstar.rs
@@ -4,6 +4,8 @@ use crate::object::RTreeObject;
 use crate::params::{InsertionStrategy, RTreeParams};
 use crate::point::{Point, PointExt};
 use crate::rtree::RTree;
+
+use alloc::vec::Vec;
 use num_traits::{Bounded, Zero};
 
 /// Inserts points according to the r-star heuristic.
@@ -55,7 +57,7 @@ impl InsertionStrategy for RStarInsertionStrategy {
                 PerformSplit(node) => {
                     // The root node was split, create a new root and increase height
                     let new_root = ParentNode::new_root::<Params>();
-                    let old_root = ::std::mem::replace(tree.root_mut(), new_root);
+                    let old_root = ::core::mem::replace(tree.root_mut(), new_root);
                     let new_envelope = old_root.envelope.merged(&node.envelope());
                     let root = tree.root_mut();
                     root.envelope = new_envelope;