+++ rstar/src/lib.rs
@@ -21,6 +21,9 @@
 //!
 #![deny(missing_docs)]
 #![forbid(unsafe_code)]
+#![cfg_attr(not(test), no_std)]
+
+extern crate alloc;
 
 mod aabb;
 mod algorithm;